Move address checksum logic to NetworkRegistry.
