+++ src/utils/EntityID.ts
@@ -19,8 +19,6 @@
  */
 
 import {byteToHex} from "./B64Utils";
-import router from "@/router";
-import {NetworkRegistry} from "@/schemas/NetworkRegistry";
 
 export class EntityID {
 
@@ -64,26 +62,9 @@ export class EntityID {
         return result
     }
 
-    public static normalize(s: string, withChecksum: boolean=false): string|null {
+    public static normalize(s: string, checksum: string|null = null): string|null {
         const id = EntityID.parse(s, true)
-        return id !== null ? id.toString() + (withChecksum ? "-" + id.makeChecksum() : "") : null
-    }
-
-    public static isValid(id: string, checksum: string | null = null): boolean {
-        const entityID = EntityID.parse(id, true)
-        return (entityID != null) && (checksum === null || entityID.isValidChecksum(checksum))
-    }
-
-    public isValidChecksum(checksum: string): boolean {
-        return this.makeChecksum() == checksum
-    }
-
-    public makeChecksum(): string {
-        const ledger = router.currentRoute.value.params.network
-        const ledgerId = (ledger === NetworkRegistry.MAIN_NETWORK) ? 0
-            : (ledger === NetworkRegistry.TEST_NETWORK) ? 1 : 2
-
-        return checksum(ledgerId, this.toString())
+        return id !== null ? id.toString() + (checksum ? "-" + checksum : "") : null
     }
 
     public toString(): string {
@@ -165,71 +146,4 @@ function compareNumber(n1: number, n2: number): number {
         result = 0
     }
     return result
-}
-
-//
-// From https://github.com/hashgraph/hedera-improvement-proposal/blob/master/assets/hip-15/HIP-15-javascript.html
-//
-// Given a ledger ID and an address like "0.0.123", return a checksum like "vfmkw" . The address must be in no-checksum
-// format, with no extra characters (so not "0.0.00123" or "==0.0.123==" or "0.0.123-vfmkw"). The algorithm is defined
-// by the HIP-15 standard to be:
-//
-// a = a valid no-checksum address string, such as 0.0.123
-// d = int array for the digits of a (using 10 to represent "."), so 0.0.123 is [0,10,0,10,1,2,3]
-// h = unsigned byte array containing the ledger ID followed by 6 zero bytes
-// p3 = 26 * 26 * 26
-// p5 = 26 * 26 * 26 * 26 * 26
-// sd0 = (d[0] + d[2] + d[4] + d[6] + ...) mod 11
-// sd1 = (d[1] + d[3] + d[5] + d[7] + ...) mod 11
-// sd = (...((((d[0] * 31) + d[1]) * 31) + d[2]) * 31 + ... ) * 31 + d[d.length-1]) mod p3
-// sh = (...(((h[0] * 31) + h[1]) * 31) + h[2]) * 31 + ... ) * 31 + h[h.length-1]) mod p5
-// c = (((d.length mod 5) * 11 + sd0) * 11 + sd1) * p3 + sd + sh ) mod p5
-// cp = (c * 1000003) mod p5
-// checksum = cp, written as 5 digits in base 26, using a-z
-//
-//in ports to other languages, answer can be a string, digits an int32[] and the rest int32 (or uint32[] and uint32)
-
-function checksum(ledgerId: number, addr: string) {
-    let answer = "";
-    const d = [];      //digits with 10 for ".", so if addr == "0.0.123" then d == [0, 10, 0, 10, 1, 2, 3] *** FIX ***
-    let sd0 = 0;      //sum of even positions (mod 11)
-    let sd1 = 0;      //sum of odd positions (mod 11)
-    let sd = 0;       //weighted sum of all positions (mod p3)
-    let sh = 0;      //hash of the ledger ID
-    let cp: number;       //the checksum, as a single number
-    const p3 = 26 * 26 * 26;           //3 digits in base 26
-    const p5 = 26 * 26 * 26 * 26 * 26; //5 digits in base 26
-    const ascii_a = "a".charCodeAt(0);  //97  *** FIX ***
-    const m = 1_000_003; //min prime greater than a million. Used for the final permutation.
-    const w = 31; //sum s of digit values weights them by powers of w. Should be coprime to p5.
-
-    let id = ledgerId + "000000000000";
-    const h = []; // *** FIX ***
-    if (id.length % 2 == 1) id = "0" + id;
-    for (let i=0; i<id.length; i+=2) {  // *** FIX ***
-        h.push(parseInt(id.substr(i,2),16));
-    }
-    for (let i = 0; i < addr.length; i++) {
-        d.push(addr[i]=="." ? 10 : parseInt(addr[i],10));
-    }
-    for (let i=0; i<d.length; i++) {
-        sd = (w * sd + d[i]) % p3;
-        if (i % 2 == 0) {
-            sd0 = (sd0 + d[i]) % 11;
-        } else {
-            sd1 = (sd1 + d[i]) % 11;
-        }
-    }
-    for (let i=0; i<h.length; i++) {
-        sh = (w * sh + h[i]) % p5;
-    }
-    const c = ((((addr.length % 5) * 11 + sd0) * 11 + sd1) * p3 + sd + sh) % p5;  //the checksum, before the final permutation
-    cp = (c * m) % p5;
-
-    for (let i=0; i<5; i++) {
-        answer = String.fromCharCode(ascii_a + (cp % 26)) + answer;
-        cp /= 26;
-    }
-
-    return answer;
 }
